% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragonking.R
\docType{package}
\name{dragonking}
\alias{dragonking}
\alias{dragonking-package}
\title{dragonking: Statistical tools for identifying dragon kings}
\description{
This package provide statistical methods to identify events in a
dataset that are dragon kings (DKs). The statistical methods in this
package were reviewed in: Wheatley S, Sornette D (2015). Multiple
outlier detection in samples with exponential & pareto tails: Redeeming
the inward approach & detecting dragon kings. Swiss Finance Institute
Research Paper Series No. 15-28.
}
