% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stats.R
\name{mrs_stat}
\alias{mrs_stat}
\title{Max-robust-sum (MRS) test statistic to identify dragon kings (DKs)}
\usage{
mrs_stat(vals, r, m)
}
\arguments{
\item{vals}{numeric vector with at least 3 elements}

\item{r}{integer indicating number of DKs in \code{vals}}

\item{m}{pre-specified maximum number of DKs in \code{vals}}
}
\value{
MRS test statistic
}
\description{
\code{mrs_stat} calculates the MRS test statistic to determine whether
there is significant support for the existence of \code{r} DKs in
\code{vals}. This test avoids denominator masking.
}
\details{
For more information, see:

Wheatley S, Sornette D (2015). Multiple outlier detection in samples with
exponential & pareto tails: Redeeming the inward approach & detecting
dragon kings. Swiss Finance Institute Research Paper Series No. 15-28. doi:
10.2139/ssrn.2645709
}
\examples{
# generate a numeric vector with DKs
temp <- c(rexp(100),   # exponentially distributed RV
          15, 15, 15)  # DK elements

# calculate test statistic for DKs
mrs_stat(temp, r = 2, m = 3)
}
